Replace HelloWorld component with RobotBuilder component
The HelloWorld.vue file was deleted as it was only a placeholder. Created and added the RobotBuilder.vue component instead for the main focus of this project, which is to build a robot. Also, updated App.vue to reflect this change, in additional to theming tweaks for a better presentation of the new component.